Created agfuse module and completed createProperty() function
Created the agfuse module which is intended to contain many the agro
fusion functionality.  The createProperty() function which creates a
shapefile from input variables was written and appears to work
correctly.  The pseudo code for isoRasterProp() has been written and
will be implemented next.  Also some test scripts have been written to
test functionality as well as prototype ideas.